ai_clone/requirements.txt
beautifulsoup4==4.13.4
chromadb==0.6.3
fastapi==0.115.8
langchain==0.3.19
langchain-chroma==0.2.2
langchain-huggingface==0.1.2
langchain-openai==0.3.6
langchain-text-splitters==0.3.6
pydantic==2.10.6
python-dotenv==1.0.1
sentence-transformers==3.4.1
streamlit==1.43.0
uvicorn==0.34.0
